Move collection condition to dedicated file
